+++ src/main/java/org/broadinstitute/consent/http/models/dataset_registration_v1/DatasetRegistrationSchemaV1.java
@@ -48,6 +48,7 @@ import java.util.Map;
     "nihICsSupportingStudy",
     "nihProgramOfficerName",
     "nihInstitutionCenterSubmission",
+    "nihInstitutionalCertificationFileName",
     "nihGenomicProgramAdministratorName",
     "multiCenterStudy",
     "collaboratingSites",
@@ -153,6 +154,7 @@ public class DatasetRegistrationSchemaV1 {
      *
      */
     @JsonProperty("nihAnvilUse")
+    @JsonPropertyDescription("NIH Anvil Use")
     private List<NihAnvilUse> nihAnvilUse = new ArrayList<NihAnvilUse>();
     /**
      * Are you planning to submit to AnVIL?
@@ -216,7 +218,7 @@ public class DatasetRegistrationSchemaV1 {
      */
     @JsonProperty("nihICsSupportingStudy")
     @JsonPropertyDescription("NIH ICs Supporting the Study")
-    private DatasetRegistrationSchemaV1 .NihICsSupportingStudy nihICsSupportingStudy;
+    private List<NihICsSupportingStudy> nihICsSupportingStudy = new ArrayList<NihICsSupportingStudy>();
     /**
      * NIH Program Officer Name
      *
@@ -231,6 +233,13 @@ public class DatasetRegistrationSchemaV1 {
     @JsonProperty("nihInstitutionCenterSubmission")
     @JsonPropertyDescription("NIH Institution/Center for Submission")
     private DatasetRegistrationSchemaV1 .NihInstitutionCenterSubmission nihInstitutionCenterSubmission;
+    /**
+     * If an Institutional Certification for this consent group exists, please upload it here (file upload)
+     *
+     */
+    @JsonProperty("nihInstitutionalCertificationFileName")
+    @JsonPropertyDescription("If an Institutional Certification for this consent group exists, please upload it here (file upload)")
+    private String nihInstitutionalCertificationFileName;
     /**
      * NIH Genomic Program Administrator Name
      *
@@ -546,11 +555,19 @@ public class DatasetRegistrationSchemaV1 {
         this.publicVisibility = publicVisibility;
     }
 
+    /**
+     * NIH Anvil Use
+     *
+     */
     @JsonProperty("nihAnvilUse")
     public List<NihAnvilUse> getNihAnvilUse() {
         return nihAnvilUse;
     }
 
+    /**
+     * NIH Anvil Use
+     *
+     */
     @JsonProperty("nihAnvilUse")
     public void setNihAnvilUse(List<NihAnvilUse> nihAnvilUse) {
         this.nihAnvilUse = nihAnvilUse;
@@ -705,7 +722,7 @@ public class DatasetRegistrationSchemaV1 {
      *
      */
     @JsonProperty("nihICsSupportingStudy")
-    public DatasetRegistrationSchemaV1 .NihICsSupportingStudy getNihICsSupportingStudy() {
+    public List<NihICsSupportingStudy> getNihICsSupportingStudy() {
         return nihICsSupportingStudy;
     }
 
@@ -714,7 +731,7 @@ public class DatasetRegistrationSchemaV1 {
      *
      */
     @JsonProperty("nihICsSupportingStudy")
-    public void setNihICsSupportingStudy(DatasetRegistrationSchemaV1 .NihICsSupportingStudy nihICsSupportingStudy) {
+    public void setNihICsSupportingStudy(List<NihICsSupportingStudy> nihICsSupportingStudy) {
         this.nihICsSupportingStudy = nihICsSupportingStudy;
     }
 
@@ -752,6 +769,24 @@ public class DatasetRegistrationSchemaV1 {
     @JsonProperty("nihInstitutionCenterSubmission")
     public void setNihInstitutionCenterSubmission(DatasetRegistrationSchemaV1 .NihInstitutionCenterSubmission nihInstitutionCenterSubmission) {
         this.nihInstitutionCenterSubmission = nihInstitutionCenterSubmission;
+    }
+
+    /**
+     * If an Institutional Certification for this consent group exists, please upload it here (file upload)
+     *
+     */
+    @JsonProperty("nihInstitutionalCertificationFileName")
+    public String getNihInstitutionalCertificationFileName() {
+        return nihInstitutionalCertificationFileName;
+    }
+
+    /**
+     * If an Institutional Certification for this consent group exists, please upload it here (file upload)
+     *
+     */
+    @JsonProperty("nihInstitutionalCertificationFileName")
+    public void setNihInstitutionalCertificationFileName(String nihInstitutionalCertificationFileName) {
+        this.nihInstitutionalCertificationFileName = nihInstitutionalCertificationFileName;
     }
 
     /**
@@ -1138,6 +1173,10 @@ public class DatasetRegistrationSchemaV1 {
         sb.append('=');
         sb.append(((this.nihInstitutionCenterSubmission == null)?"<null>":this.nihInstitutionCenterSubmission));
         sb.append(',');
+        sb.append("nihInstitutionalCertificationFileName");
+        sb.append('=');
+        sb.append(((this.nihInstitutionalCertificationFileName == null)?"<null>":this.nihInstitutionalCertificationFileName));
+        sb.append(',');
         sb.append("nihGenomicProgramAdministratorName");
         sb.append('=');
         sb.append(((this.nihGenomicProgramAdministratorName == null)?"<null>":this.nihGenomicProgramAdministratorName));
@@ -1230,6 +1269,7 @@ public class DatasetRegistrationSchemaV1 {
         result = ((result* 31)+((this.alternativeDataSharingPlanTargetPublicReleaseDate == null)? 0 :this.alternativeDataSharingPlanTargetPublicReleaseDate.hashCode()));
         result = ((result* 31)+((this.nihGrantContractNumber == null)? 0 :this.nihGrantContractNumber.hashCode()));
         result = ((result* 31)+((this.studyType == null)? 0 :this.studyType.hashCode()));
+        result = ((result* 31)+((this.nihInstitutionalCertificationFileName == null)? 0 :this.nihInstitutionalCertificationFileName.hashCode()));
         result = ((result* 31)+((this.phenotypeIndication == null)? 0 :this.phenotypeIndication.hashCode()));
         result = ((result* 31)+((this.alternativeDataSharingPlanControlledOpenAccess == null)? 0 :this.alternativeDataSharingPlanControlledOpenAccess.hashCode()));
         result = ((result* 31)+((this.sequencingCenter == null)? 0 :this.sequencingCenter.hashCode()));
@@ -1263,7 +1303,7 @@ public class DatasetRegistrationSchemaV1 {
             return false;
         }
         DatasetRegistrationSchemaV1 rhs = ((DatasetRegistrationSchemaV1) other);
-        return ((((((((((((((((((((((((((((((((((((((((this.nihGenomicProgramAdministratorName == rhs.nihGenomicProgramAdministratorName)||((this.nihGenomicProgramAdministratorName!= null)&&this.nihGenomicProgramAdministratorName.equals(rhs.nihGenomicProgramAdministratorName)))&&((this.alternativeDataSharingPlanFileName == rhs.alternativeDataSharingPlanFileName)||((this.alternativeDataSharingPlanFileName!= null)&&this.alternativeDataSharingPlanFileName.equals(rhs.alternativeDataSharingPlanFileName))))&&((this.studyName == rhs.studyName)||((this.studyName!= null)&&this.studyName.equals(rhs.studyName))))&&((this.piEmail == rhs.piEmail)||((this.piEmail!= null)&&this.piEmail.equals(rhs.piEmail))))&&((this.dataSubmitterUserId == rhs.dataSubmitterUserId)||((this.dataSubmitterUserId!= null)&&this.dataSubmitterUserId.equals(rhs.dataSubmitterUserId))))&&((this.publicVisibility == rhs.publicVisibility)||((this.publicVisibility!= null)&&this.publicVisibility.equals(rhs.publicVisibility))))&&((this.dataCustodianEmail == rhs.dataCustodianEmail)||((this.dataCustodianEmail!= null)&&this.dataCustodianEmail.equals(rhs.dataCustodianEmail))))&&((this.alternativeDataSharingPlanDataSubmitted == rhs.alternativeDataSharingPlanDataSubmitted)||((this.alternativeDataSharingPlanDataSubmitted!= null)&&this.alternativeDataSharingPlanDataSubmitted.equals(rhs.alternativeDataSharingPlanDataSubmitted))))&&((this.submittingToAnvil == rhs.submittingToAnvil)||((this.submittingToAnvil!= null)&&this.submittingToAnvil.equals(rhs.submittingToAnvil))))&&((this.dbGaPStudyRegistrationName == rhs.dbGaPStudyRegistrationName)||((this.dbGaPStudyRegistrationName!= null)&&this.dbGaPStudyRegistrationName.equals(rhs.dbGaPStudyRegistrationName))))&&((this.collaboratingSites == rhs.collaboratingSites)||((this.collaboratingSites!= null)&&this.collaboratingSites.equals(rhs.collaboratingSites))))&&((this.alternativeDataSharingPlanTargetDeliveryDate == rhs.alternativeDataSharingPlanTargetDeliveryDate)||((this.alternativeDataSharingPlanTargetDeliveryDate!= null)&&this.alternativeDataSharingPlanTargetDeliveryDate.equals(rhs.alternativeDataSharingPlanTargetDeliveryDate))))&&((this.nihProgramOfficerName == rhs.nihProgramOfficerName)||((this.nihProgramOfficerName!= null)&&this.nihProgramOfficerName.equals(rhs.nihProgramOfficerName))))&&((this.dataTypes == rhs.dataTypes)||((this.dataTypes!= null)&&this.dataTypes.equals(rhs.dataTypes))))&&((this.alternativeDataSharingPlanTargetPublicReleaseDate == rhs.alternativeDataSharingPlanTargetPublicReleaseDate)||((this.alternativeDataSharingPlanTargetPublicReleaseDate!= null)&&this.alternativeDataSharingPlanTargetPublicReleaseDate.equals(rhs.alternativeDataSharingPlanTargetPublicReleaseDate))))&&((this.nihGrantContractNumber == rhs.nihGrantContractNumber)||((this.nihGrantContractNumber!= null)&&this.nihGrantContractNumber.equals(rhs.nihGrantContractNumber))))&&((this.studyType == rhs.studyType)||((this.studyType!= null)&&this.studyType.equals(rhs.studyType))))&&((this.phenotypeIndication == rhs.phenotypeIndication)||((this.phenotypeIndication!= null)&&this.phenotypeIndication.equals(rhs.phenotypeIndication))))&&((this.alternativeDataSharingPlanControlledOpenAccess == rhs.alternativeDataSharingPlanControlledOpenAccess)||((this.alternativeDataSharingPlanControlledOpenAccess!= null)&&this.alternativeDataSharingPlanControlledOpenAccess.equals(rhs.alternativeDataSharingPlanControlledOpenAccess))))&&((this.sequencingCenter == rhs.sequencingCenter)||((this.sequencingCenter!= null)&&this.sequencingCenter.equals(rhs.sequencingCenter))))&&((this.multiCenterStudy == rhs.multiCenterStudy)||((this.multiCenterStudy!= null)&&this.multiCenterStudy.equals(rhs.multiCenterStudy))))&&((this.alternativeDataSharingPlanDataReleased == rhs.alternativeDataSharingPlanDataReleased)||((this.alternativeDataSharingPlanDataReleased!= null)&&this.alternativeDataSharingPlanDataReleased.equals(rhs.alternativeDataSharingPlanDataReleased))))&&((this.consentGroups == rhs.consentGroups)||((this.consentGroups!= null)&&this.consentGroups.equals(rhs.consentGroups))))&&((this.studyDescription == rhs.studyDescription)||((this.studyDescription!= null)&&this.studyDescription.equals(rhs.studyDescription))))&&((this.nihAnvilUse == rhs.nihAnvilUse)||((this.nihAnvilUse!= null)&&this.nihAnvilUse.equals(rhs.nihAnvilUse))))&&((this.dataAccessCommitteeId == rhs.dataAccessCommitteeId)||((this.dataAccessCommitteeId!= null)&&this.dataAccessCommitteeId.equals(rhs.dataAccessCommitteeId))))&&((this.nihICsSupportingStudy == rhs.nihICsSupportingStudy)||((this.nihICsSupportingStudy!= null)&&this.nihICsSupportingStudy.equals(rhs.nihICsSupportingStudy)))))&&((this.piName == rhs.piName)||((this.piName!= null)&&this.piName.equals(rhs.piName))))&&((this.dbGaPPhsID == rhs.dbGaPPhsID)||((this.dbGaPPhsID!= null)&&this.dbGaPPhsID.equals(rhs.dbGaPPhsID))))&&((this.alternativeDataSharingPlan == rhs.alternativeDataSharingPlan)||((this.alternativeDataSharingPlan!= null)&&this.alternativeDataSharingPlan.equals(rhs.alternativeDataSharingPlan))))&&((this.alternativeDataSharingPlanReasons == rhs.alternativeDataSharingPlanReasons)||((this.alternativeDataSharingPlanReasons!= null)&&this.alternativeDataSharingPlanReasons.equals(rhs.alternativeDataSharingPlanReasons))))&&((this.species == rhs.species)||((this.species!= null)&&this.species.equals(rhs.species))))&&((this.embargoReleaseDate == rhs.embargoReleaseDate)||((this.embargoReleaseDate!= null)&&this.embargoReleaseDate.equals(rhs.embargoReleaseDate))))&&((this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation == rhs.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation)||((this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation!= null)&&this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation.equals(rhs.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation))))&&((this.alternativeDataSharingPlanExplanation == rhs.alternativeDataSharingPlanExplanation)||((this.alternativeDataSharingPlanExplanation!= null)&&this.alternativeDataSharingPlanExplanation.equals(rhs.alternativeDataSharingPlanExplanation))))&&((this.nihInstitutionCenterSubmission == rhs.nihInstitutionCenterSubmission)||((this.nihInstitutionCenterSubmission!= null)&&this.nihInstitutionCenterSubmission.equals(rhs.nihInstitutionCenterSubmission))))&&((this.controlledAccessRequiredForGenomicSummaryResultsGSR == rhs.controlledAccessRequiredForGenomicSummaryResultsGSR)||((this.controlledAccessRequiredForGenomicSummaryResultsGSR!= null)&&this.controlledAccessRequiredForGenomicSummaryResultsGSR.equals(rhs.controlledAccessRequiredForGenomicSummaryResultsGSR))))&&((this.piInstitution == rhs.piInstitution)||((this.piInstitution!= null)&&this.piInstitution.equals(rhs.piInstitution))));
+        return ((((((((((((((((((((((((((((((((((((((((this.nihGenomicProgramAdministratorName == rhs.nihGenomicProgramAdministratorName)||((this.nihGenomicProgramAdministratorName!= null)&&this.nihGenomicProgramAdministratorName.equals(rhs.nihGenomicProgramAdministratorName)))&&((this.alternativeDataSharingPlanFileName == rhs.alternativeDataSharingPlanFileName)||((this.alternativeDataSharingPlanFileName!= null)&&this.alternativeDataSharingPlanFileName.equals(rhs.alternativeDataSharingPlanFileName))))&&((this.studyName == rhs.studyName)||((this.studyName!= null)&&this.studyName.equals(rhs.studyName))))&&((this.piEmail == rhs.piEmail)||((this.piEmail!= null)&&this.piEmail.equals(rhs.piEmail))))&&((this.dataSubmitterUserId == rhs.dataSubmitterUserId)||((this.dataSubmitterUserId!= null)&&this.dataSubmitterUserId.equals(rhs.dataSubmitterUserId))))&&((this.publicVisibility == rhs.publicVisibility)||((this.publicVisibility!= null)&&this.publicVisibility.equals(rhs.publicVisibility))))&&((this.dataCustodianEmail == rhs.dataCustodianEmail)||((this.dataCustodianEmail!= null)&&this.dataCustodianEmail.equals(rhs.dataCustodianEmail))))&&((this.alternativeDataSharingPlanDataSubmitted == rhs.alternativeDataSharingPlanDataSubmitted)||((this.alternativeDataSharingPlanDataSubmitted!= null)&&this.alternativeDataSharingPlanDataSubmitted.equals(rhs.alternativeDataSharingPlanDataSubmitted))))&&((this.submittingToAnvil == rhs.submittingToAnvil)||((this.submittingToAnvil!= null)&&this.submittingToAnvil.equals(rhs.submittingToAnvil))))&&((this.dbGaPStudyRegistrationName == rhs.dbGaPStudyRegistrationName)||((this.dbGaPStudyRegistrationName!= null)&&this.dbGaPStudyRegistrationName.equals(rhs.dbGaPStudyRegistrationName))))&&((this.collaboratingSites == rhs.collaboratingSites)||((this.collaboratingSites!= null)&&this.collaboratingSites.equals(rhs.collaboratingSites))))&&((this.alternativeDataSharingPlanTargetDeliveryDate == rhs.alternativeDataSharingPlanTargetDeliveryDate)||((this.alternativeDataSharingPlanTargetDeliveryDate!= null)&&this.alternativeDataSharingPlanTargetDeliveryDate.equals(rhs.alternativeDataSharingPlanTargetDeliveryDate))))&&((this.nihProgramOfficerName == rhs.nihProgramOfficerName)||((this.nihProgramOfficerName!= null)&&this.nihProgramOfficerName.equals(rhs.nihProgramOfficerName))))&&((this.dataTypes == rhs.dataTypes)||((this.dataTypes!= null)&&this.dataTypes.equals(rhs.dataTypes))))&&((this.alternativeDataSharingPlanTargetPublicReleaseDate == rhs.alternativeDataSharingPlanTargetPublicReleaseDate)||((this.alternativeDataSharingPlanTargetPublicReleaseDate!= null)&&this.alternativeDataSharingPlanTargetPublicReleaseDate.equals(rhs.alternativeDataSharingPlanTargetPublicReleaseDate))))&&((this.nihGrantContractNumber == rhs.nihGrantContractNumber)||((this.nihGrantContractNumber!= null)&&this.nihGrantContractNumber.equals(rhs.nihGrantContractNumber))))&&((this.studyType == rhs.studyType)||((this.studyType!= null)&&this.studyType.equals(rhs.studyType))))&&((this.nihInstitutionalCertificationFileName == rhs.nihInstitutionalCertificationFileName)||((this.nihInstitutionalCertificationFileName!= null)&&this.nihInstitutionalCertificationFileName.equals(rhs.nihInstitutionalCertificationFileName))))&&((this.phenotypeIndication == rhs.phenotypeIndication)||((this.phenotypeIndication!= null)&&this.phenotypeIndication.equals(rhs.phenotypeIndication))))&&((this.alternativeDataSharingPlanControlledOpenAccess == rhs.alternativeDataSharingPlanControlledOpenAccess)||((this.alternativeDataSharingPlanControlledOpenAccess!= null)&&this.alternativeDataSharingPlanControlledOpenAccess.equals(rhs.alternativeDataSharingPlanControlledOpenAccess))))&&((this.sequencingCenter == rhs.sequencingCenter)||((this.sequencingCenter!= null)&&this.sequencingCenter.equals(rhs.sequencingCenter))))&&((this.multiCenterStudy == rhs.multiCenterStudy)||((this.multiCenterStudy!= null)&&this.multiCenterStudy.equals(rhs.multiCenterStudy))))&&((this.alternativeDataSharingPlanDataReleased == rhs.alternativeDataSharingPlanDataReleased)||((this.alternativeDataSharingPlanDataReleased!= null)&&this.alternativeDataSharingPlanDataReleased.equals(rhs.alternativeDataSharingPlanDataReleased))))&&((this.consentGroups == rhs.consentGroups)||((this.consentGroups!= null)&&this.consentGroups.equals(rhs.consentGroups))))&&((this.studyDescription == rhs.studyDescription)||((this.studyDescription!= null)&&this.studyDescription.equals(rhs.studyDescription))))&&((this.nihAnvilUse == rhs.nihAnvilUse)||((this.nihAnvilUse!= null)&&this.nihAnvilUse.equals(rhs.nihAnvilUse))))&&((this.dataAccessCommitteeId == rhs.dataAccessCommitteeId)||((this.dataAccessCommitteeId!= null)&&this.dataAccessCommitteeId.equals(rhs.dataAccessCommitteeId))))&&((this.nihICsSupportingStudy == rhs.nihICsSupportingStudy)||((this.nihICsSupportingStudy!= null)&&this.nihICsSupportingStudy.equals(rhs.nihICsSupportingStudy))))&&((this.piName == rhs.piName)||((this.piName!= null)&&this.piName.equals(rhs.piName))))&&((this.dbGaPPhsID == rhs.dbGaPPhsID)||((this.dbGaPPhsID!= null)&&this.dbGaPPhsID.equals(rhs.dbGaPPhsID))))&&((this.alternativeDataSharingPlan == rhs.alternativeDataSharingPlan)||((this.alternativeDataSharingPlan!= null)&&this.alternativeDataSharingPlan.equals(rhs.alternativeDataSharingPlan))))&&((this.alternativeDataSharingPlanReasons == rhs.alternativeDataSharingPlanReasons)||((this.alternativeDataSharingPlanReasons!= null)&&this.alternativeDataSharingPlanReasons.equals(rhs.alternativeDataSharingPlanReasons))))&&((this.species == rhs.species)||((this.species!= null)&&this.species.equals(rhs.species))))&&((this.embargoReleaseDate == rhs.embargoReleaseDate)||((this.embargoReleaseDate!= null)&&this.embargoReleaseDate.equals(rhs.embargoReleaseDate))))&&((this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation == rhs.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation)||((this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation!= null)&&this.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation.equals(rhs.controlledAccessRequiredForGenomicSummaryResultsGSRNotRequiredExplanation))))&&((this.alternativeDataSharingPlanExplanation == rhs.alternativeDataSharingPlanExplanation)||((this.alternativeDataSharingPlanExplanation!= null)&&this.alternativeDataSharingPlanExplanation.equals(rhs.alternativeDataSharingPlanExplanation))))&&((this.nihInstitutionCenterSubmission == rhs.nihInstitutionCenterSubmission)||((this.nihInstitutionCenterSubmission!= null)&&this.nihInstitutionCenterSubmission.equals(rhs.nihInstitutionCenterSubmission))))&&((this.controlledAccessRequiredForGenomicSummaryResultsGSR == rhs.controlledAccessRequiredForGenomicSummaryResultsGSR)||((this.controlledAccessRequiredForGenomicSummaryResultsGSR!= null)&&this.controlledAccessRequiredForGenomicSummaryResultsGSR.equals(rhs.controlledAccessRequiredForGenomicSummaryResultsGSR))))&&((this.piInstitution == rhs.piInstitution)||((this.piInstitution!= null)&&this.piInstitution.equals(rhs.piInstitution))));
     }
 
 
@@ -1347,76 +1387,6 @@ public class DatasetRegistrationSchemaV1 {
         @JsonCreator
         public static DatasetRegistrationSchemaV1 .AlternativeDataSharingPlanDataSubmitted fromValue(String value) {
             DatasetRegistrationSchemaV1 .AlternativeDataSharingPlanDataSubmitted constant = CONSTANTS.get(value);
-            if (constant == null) {
-                throw new IllegalArgumentException(value);
-            } else {
-                return constant;
-            }
-        }
-
-    }
-
-
-    /**
-     * NIH ICs Supporting the Study
-     *
-     */
-    @Generated("jsonschema2pojo")
-    public enum NihICsSupportingStudy {
-
-        NCI("NCI"),
-        NEI("NEI"),
-        NHLBI("NHLBI"),
-        NHGRI("NHGRI"),
-        NIA("NIA"),
-        NIAAA("NIAAA"),
-        NIAID("NIAID"),
-        NIAMS("NIAMS"),
-        NIBIB("NIBIB"),
-        NICHD("NICHD"),
-        NIDCD("NIDCD"),
-        NIDCR("NIDCR"),
-        NIDDK("NIDDK"),
-        NIDA("NIDA"),
-        NIEHS("NIEHS"),
-        NIGMS("NIGMS"),
-        NIMH("NIMH"),
-        NIMHD("NIMHD"),
-        NINDS("NINDS"),
-        NINR("NINR"),
-        NLM("NLM"),
-        CC("CC"),
-        CIT("CIT"),
-        CSR("CSR"),
-        FIC("FIC"),
-        NCATS("NCATS"),
-        NCCIH("NCCIH");
-        private final String value;
-        private final static Map<String, DatasetRegistrationSchemaV1 .NihICsSupportingStudy> CONSTANTS = new HashMap<String, DatasetRegistrationSchemaV1 .NihICsSupportingStudy>();
-
-        static {
-            for (DatasetRegistrationSchemaV1 .NihICsSupportingStudy c: values()) {
-                CONSTANTS.put(c.value, c);
-            }
-        }
-
-        NihICsSupportingStudy(String value) {
-            this.value = value;
-        }
-
-        @Override
-        public String toString() {
-            return this.value;
-        }
-
-        @JsonValue
-        public String value() {
-            return this.value;
-        }
-
-        @JsonCreator
-        public static DatasetRegistrationSchemaV1 .NihICsSupportingStudy fromValue(String value) {
-            DatasetRegistrationSchemaV1 .NihICsSupportingStudy constant = CONSTANTS.get(value);
             if (constant == null) {
                 throw new IllegalArgumentException(value);
             } else {
